requirements.txt
slack_sdk
slack-bolt
flask
wheel
uwsgi
python-dotenv
